Better next/previous buffers switching
The current spacemacs-next/prev-real-buffer check the start of a buffer
to see if it contains a "*" character and ignore the buffer if it sees
one. This makes valid buffers, such as `*eshell*` or `*Backtrace*` or
`*Message*` being ignored.

We better let users choose which to ignore or include, and provide a
default for using spacemacs-boring-buffer-regexp-list and
spacemacs-interesting-buffers, and add a function that uses these two
lists to decide whether a buffer should be skipped or not.

Also bind the spacemacs-next/prev-real-buffer commands to "bn" and "bp"
respectively.